# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Studio Hub Architects API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: Services list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Team ────────────────────────────────────────────────────────────────────
  /team:
    get:
      operationId: listTeamMembers
      tags: [team]
      summary: List all team members
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createTeamMember
      tags: [team]
      summary: Create a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "201":
          description: Created member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/{id}:
    patch:
      operationId: updateTeamMember
      tags: [team]
      summary: Update a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: deleteTeamMember
      tags: [team]
      summary: Delete a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Testimonials ─────────────────────────────────────────────────────────────
  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List testimonials
      responses:
        "200":
          description: Testimonials list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Create a testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "201":
          description: Created testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /testimonials/{id}:
    patch:
      operationId: updateTestimonial
      tags: [testimonials]
      summary: Update a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialUpdate"
      responses:
        "200":
          description: Updated testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
    delete:
      operationId: deleteTestimonial
      tags: [testimonials]
      summary: Delete a testimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Blog ─────────────────────────────────────────────────────────────────────
...
  /milestones:
    get:
      operationId: listMilestones
      tags: [milestones]
      summary: List milestones
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Milestones list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Milestone"
    post:
      operationId: createMilestone
      tags: [milestones]
      summary: Create a project milestone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneInput"
      responses:
        "201":
          description: Created milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
  /milestones/{id}:
    patch:
      operationId: updateMilestone
      tags: [milestones]
      summary: Update a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneUpdate"
      responses:
        "200":
          description: Updated milestone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
    delete:
      operationId: deleteMilestone
      tags: [milestones]
      summary: Delete a milestone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Settings ─────────────────────────────────────────────────────────────────
...
  /settings/{key}:
    put:
      operationId: upsertSetting
      tags: [settings]
      summary: Create or update a site setting
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingInput"
      responses:
        "200":
          description: Updated setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setting"

  # ── Analytics ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, title, category, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: ["string", "null"]
        category:
          type: string
          description: "Residential | Hospitality | Commercial | Site Supervision"
        description:
          type: ["string", "null"]
        story:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        area:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        status:
          type: string
          description: "planning | in_progress | completed"
        featured:
          type: boolean
        coverImage:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        floorPlans:
          type: array
          items:
            type: string
        videos:
          type: array
          items:
            type: string
        materials:
          type: array
          items:
            type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        clientName:
          type: ["string", "null"]
        budget:
          type: ["string", "null"]
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [title, category, status]
      properties:
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        description:
          type: string
        story:
          type: string
        location:
          type: string
        area:
          type: string
        year:
          type: integer
        status:
          type: string
        featured:
          type: boolean
        coverImage:
          type: string
        images:
          type: array
          items:
            type: string
        floorPlans:
          type: array
          items:
            type: string
        videos:
          type: array
          items:
            type: string
        materials:
          type: array
          items:
            type: string
        latitude:
          type: number
        longitude:
          type: number
        clientName:
          type: string
        budget:
          type: string
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        category:
          type: string
        description:
          type: string
        story:
          type: string
        location:
          type: string
        area:
          type: string
        year:
          type: integer
        status:
          type: string
        featured:
          type: boolean
        coverImage:
          type: string
        images:
          type: array
          items:
            type: string
        floorPlans:
          type: array
          items:
            type: string
        videos:
          type: array
          items:
            type: string
        materials:
          type: array
          items:
            type: string
        latitude:
          type: number
        longitude:
          type: number
        clientName:
          type: string
        budget:
          type: string
...
    Service:
      type: object
      required: [id, title, description, sortOrder]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        icon:
          type: ["string", "null"]
        illustration:
          type: ["string", "null"]
        sortOrder:
          type: integer
    ServiceInput:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        illustration:
          type: string
        sortOrder:
          type: integer
    ServiceUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        illustration:
          type: string
        sortOrder:
          type: integer

    # ── Team ─────────────────────────────────────────────────────────────────────
    TeamMember:
      type: object
      required: [id, name, role, sortOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        bio:
          type: ["string", "null"]
        photo:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        linkedin:
          type: ["string", "null"]
        sortOrder:
          type: integer
    TeamMemberInput:
      type: object
      required: [name, role]
      properties:
        name:
          type: string
        role:
          type: string
        bio:
          type: string
        photo:
          type: string
        email:
          type: string
        linkedin:
          type: string
        sortOrder:
          type: integer
    TeamMemberUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        bio:
          type: string
        photo:
          type: string
        email:
          type: string
        linkedin:
          type: string
        sortOrder:
          type: integer

    # ── Testimonial ───────────────────────────────────────────────────────────────
    Testimonial:
      type: object
      required: [id, clientName, content, rating, createdAt]
      properties:
        id:
          type: integer
        clientName:
          type: string
        company:
          type: ["string", "null"]
        content:
          type: string
        rating:
          type: integer
        photo:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        featured:
          type: boolean
        createdAt:
          type: string
    TestimonialInput:
      type: object
      required: [clientName, content, rating]
      properties:
        clientName:
          type: string
        company:
          type: string
        content:
          type: string
        rating:
          type: integer
        photo:
          type: string
        projectId:
          type: integer
        featured:
          type: boolean
    TestimonialUpdate:
      type: object
      properties:
        clientName:
          type: string
        company:
          type: string
        content:
          type: string
        rating:
          type: integer
        photo:
          type: string
        projectId:
          type: integer
        featured:
          type: boolean

    # ── Blog ─────────────────────────────────────────────────────────────────────
...
    Milestone:
      type: object
      required: [id, projectId, title, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          description: "pending | in_progress | completed"
        dueDate:
          type: ["string", "null"]
        completedDate:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
    MilestoneInput:
      type: object
      required: [projectId, title, status]
      properties:
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
        dueDate:
          type: string
        completedDate:
          type: string
        sortOrder:
          type: integer
    MilestoneUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        dueDate:
          type: string
        completedDate:
          type: string
        sortOrder:
          type: integer

    # ── Setting ───────────────────────────────────────────────────────────────────
    Setting:
      type: object
      required: [id, key, value]
      properties:
        id:
          type: integer
        key:
          type: string
        value:
          type: string
        updatedAt:
          type: string
    SettingInput:
      type: object
      required: [value]
      properties:
        value:
          type: string

    # ── Analytics ─────────────────────────────────────────────────────────────────
...