# lib/api-zod/src/generated/types/serviceUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Studio Hub Architects API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceUpdate {
  title?: string;
  description?: string;
  icon?: string;
  illustration?: string;
  sortOrder?: number;
}