# lib/api-zod/src/generated/types/teamMember.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Studio Hub Architects API
 * OpenAPI spec version: 0.1.0
 */

export interface TeamMember {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  photo?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  linkedin?: string | null;
  sortOrder: number;
}